Add a color-independent saturation-based highlighting algo
The original algorithm for detecting highlighted areas used a
certain range of HSV values.  A new algorithm, accessible using the
method ``sat_extract()``, detects areas of high saturation.  As
highlighter colors generally have high saturation, and saturation
is independent of color, it is anticipated this algorithm will work
for a variety of highlighter colors.

However, there is still an issue, which is that areas off of the page
may be high-saturation areas as well.  A further update is needed
to avoid this problem.